Fix Styles, Done Tasks and UseLocalStorage Component
