+++ arquivos/index.css
@@ -1,1 +1,1 @@
-body{margin:0px;font-family:-apple-system,BlinkMacSystemFont,"Segoe UI","Roboto","Oxygen","Ubuntu","Cantarell","Fira Sans","Droid Sans","Helvetica Neue",sans-serif;-webkit-font-smoothing:antialiased;-moz-osx-font-smoothing:grayscale;background:#020914;color:#e7edf3;width:100vw}code{font-family:source-code-pro,Menlo,Monaco,Consolas,"Courier New",monospace}#root{margin:0px 24px;position:relative;top:-1px;min-height:100vh}.CreateTodoButton{background-color:#53a0f8;box-shadow:0px 5px 25px rgba(24,68,140,.5);border:none;border-radius:50%;cursor:pointer;font-size:50px;position:fixed;bottom:24px;right:24px;font-weight:bold;color:#fafafa;display:flex;justify-content:center;align-items:center;height:64px;width:64px;transform:rotate(0);transition:.3s ease}.CreateTodoButton:hover{transform:rotate(224deg);background-color:#f85353}.to-do-counter{font-size:1.8rem;text-align:center;margin:0;padding:30px;font-weight:700}.to-do-counter span{color:#53a0f8}.ToDoItem{background:#161b22;color:#f0f6fc;position:relative;display:flex;justify-content:center;align-items:center;margin-top:24px;box-shadow:0px 5px 50px rgba(32,35,41,.15)}.TodoItem-p{margin:14px 14px 14px 3.5rem;width:calc(100% - 50px);font-size:1rem;line-height:26px;font-weight:500}.TodoItem-p--complete{text-decoration:line-through}.Icon{cursor:pointer;display:flex;justify-content:center;align-items:center;height:48px;width:48px;font-size:1.3rem;font-weight:bold}.Icon-check{position:absolute;left:5px}.Icon-check--active{color:#53f853}.Icon-delete{position:absolute;top:-20px;right:-1.2rem}.Icon-delete:hover{color:#f85353}ul{margin:30px 0px;padding:0px 0px 56px 0px;list-style:none}.TodoSearch{font-family:-apple-system,BlinkMacSystemFont,"Segoe UI","Roboto","Oxygen","Ubuntu","Cantarell","Fira Sans","Droid Sans","Helvetica Neue",sans-serif;background:#f9fbfc;border-radius:2px;border:2px solid #202329;margin:0px 10px;height:44px;width:calc(100% - 30px);font-size:18px;text-align:center;font-weight:500;color:#1e1e1f;box-shadow:0px 5px 50px rgba(32,35,41,.25)}.TodoSearch::-moz-placeholder{color:#a5a5a5;font-weight:500;font-family:-apple-system,BlinkMacSystemFont,"Segoe UI","Roboto","Oxygen","Ubuntu","Cantarell","Fira Sans","Droid Sans","Helvetica Neue",sans-serif;font-size:14px}.TodoSearch:-ms-input-placeholder{color:#a5a5a5;font-weight:500;font-family:-apple-system,BlinkMacSystemFont,"Segoe UI","Roboto","Oxygen","Ubuntu","Cantarell","Fira Sans","Droid Sans","Helvetica Neue",sans-serif;font-size:14px}.TodoSearch::placeholder{color:#a5a5a5;font-weight:500;font-family:-apple-system,BlinkMacSystemFont,"Segoe UI","Roboto","Oxygen","Ubuntu","Cantarell","Fira Sans","Droid Sans","Helvetica Neue",sans-serif;font-size:14px}.TodoSearch:focus{outline-color:#53a0f8;box-shadow:0px 5px 25px rgba(24,68,140,.5);transition:all .5s}
+body{margin:0px;font-family:-apple-system,BlinkMacSystemFont,"Segoe UI","Roboto","Oxygen","Ubuntu","Cantarell","Fira Sans","Droid Sans","Helvetica Neue",sans-serif;-webkit-font-smoothing:antialiased;-moz-osx-font-smoothing:grayscale;background:#020914;color:#e7edf3;width:100vw;display:flex;justify-content:center}code{font-family:source-code-pro,Menlo,Monaco,Consolas,"Courier New",monospace}#root{margin:0px 24px;position:relative;top:-1px;min-height:100vh;max-width:600px}.CreateTodoButton{background-color:#53a0f8;box-shadow:0px 5px 25px rgba(24,68,140,.5);border:none;border-radius:50%;cursor:pointer;font-size:50px;position:fixed;bottom:24px;right:24px;font-weight:bold;color:#fafafa;display:flex;justify-content:center;align-items:center;height:64px;width:64px;transform:rotate(0);transition:.3s ease}.CreateTodoButton:hover{transform:rotate(224deg);background-color:#f85353}.to-do-counter{font-size:1.8rem;text-align:center;margin:0;padding:30px;font-weight:700}.to-do-counter span{color:#53a0f8}.ToDoItem{background:#161b22;color:#f0f6fc;position:relative;display:flex;justify-content:center;align-items:center;margin-top:24px;box-shadow:0px 5px 50px rgba(32,35,41,.15)}.TodoItem-p{margin:14px 14px 14px 3.5rem;width:calc(100% - 50px);font-size:1rem;line-height:26px;font-weight:500}.TodoItem-p--complete{text-decoration:line-through}.Icon{cursor:pointer;display:flex;justify-content:center;align-items:center;height:48px;width:48px;font-size:1.3rem;font-weight:bold}.Icon-check{position:absolute;left:5px}.Icon-check--active{color:#53f853}.Icon-delete{position:absolute;top:-20px;right:-1.2rem}.Icon-delete:hover{color:#f85353}ul{margin:30px 0px;padding:0px 0px 56px 0px;list-style:none}.TodoSearch{font-family:-apple-system,BlinkMacSystemFont,"Segoe UI","Roboto","Oxygen","Ubuntu","Cantarell","Fira Sans","Droid Sans","Helvetica Neue",sans-serif;background:#f9fbfc;border-radius:2px;border:2px solid #202329;margin:0px 10px;height:44px;width:calc(100% - 30px);font-size:18px;text-align:center;font-weight:500;color:#1e1e1f;box-shadow:0px 5px 50px rgba(32,35,41,.25)}.TodoSearch::-moz-placeholder{color:#a5a5a5;font-weight:500;font-family:-apple-system,BlinkMacSystemFont,"Segoe UI","Roboto","Oxygen","Ubuntu","Cantarell","Fira Sans","Droid Sans","Helvetica Neue",sans-serif;font-size:14px}.TodoSearch:-ms-input-placeholder{color:#a5a5a5;font-weight:500;font-family:-apple-system,BlinkMacSystemFont,"Segoe UI","Roboto","Oxygen","Ubuntu","Cantarell","Fira Sans","Droid Sans","Helvetica Neue",sans-serif;font-size:14px}.TodoSearch::placeholder{color:#a5a5a5;font-weight:500;font-family:-apple-system,BlinkMacSystemFont,"Segoe UI","Roboto","Oxygen","Ubuntu","Cantarell","Fira Sans","Droid Sans","Helvetica Neue",sans-serif;font-size:14px}.TodoSearch:focus{outline-color:#53a0f8;box-shadow:0px 5px 25px rgba(24,68,140,.5);transition:all .5s}
