+++ arquivos/index.js
@@ -133,7 +133,11 @@ function App() {
     const completeTodo = (text) => {
         const todoIndex = todos.findIndex(todo => todo.text === text);
         const newTodos = [...todos];
-        newTodos[todoIndex].completed = true;
+        if (newTodos[todoIndex].completed == true) {
+            newTodos[todoIndex].completed = false;
+        } else {
+            newTodos[todoIndex].completed = true;
+        }
         saveTodos(newTodos);
     };
