+++ react-app/react-app/src/index.css
@@ -7,6 +7,8 @@ body {
   -moz-osx-font-smoothing: grayscale;
   background: #020914;
   color: #e7edf3;
+  display: flex;
+  justify-content: center;
 }
 
 code {
@@ -19,4 +21,5 @@ code {
   position: relative;
   top: -1px;
   min-height: 100vh;
+  max-width: 600px;
 }
